+++ lib/src/default_cupertino_theme.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
 import 'package:x_theme_provider/src/app_theme.dart';
 
 /// This class provides a basic theme for test purpose.
-class DefaultCupertinoTheme {
+class DefaultCupertinoTheme extends AppTheme<CupertinoThemeData> {
   // ignore: public_member_api_docs
   static const CupertinoThemeData light = CupertinoThemeData(
     brightness: Brightness.light,
@@ -24,7 +23,7 @@ class DefaultCupertinoTheme {
     ),
   );
 
-  /// Returns the appropriate theme based on ThemeMode.
-  static CupertinoThemeData of(ThemeMode mode) =>
-      AppTheme.of(mode, light, dark);
+  /// Make a list of themes
+  @override
+  List<CupertinoThemeData> themeList() => super.from(light, dark,);      
 }