+++ lib/src/default_material_theme.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:x_theme_provider/src/app_theme.dart';
 
 /// This class provides a basic theme for test purpose.
-class DefaultMaterialTheme {
+class DefaultMaterialTheme extends AppTheme<ThemeData> {
   // ignore: public_member_api_docs
   static final ThemeData light = ThemeData(
     brightness: Brightness.light,
@@ -17,6 +17,7 @@ class DefaultMaterialTheme {
     useMaterial3: true,
   );
 
-  /// Returns the appropriate theme based on ThemeMode.
-  static ThemeData of(ThemeMode mode) => AppTheme.of(mode, light, dark);
+  /// Make a list of themes
+  @override
+  List<ThemeData> themeList() => super.from(light, dark,);
 }