+++ lib/src/inherited_theme_provider.dart
@@ -1,11 +1,10 @@
-import 'package:flutter/material.dart' show ThemeMode;
 import 'package:flutter/widgets.dart';
 
 /// Provides current mode to its children.
 /// This is an internal widget and should not be used directly.
 class InheritedThemeProvider extends InheritedWidget {
   /// current mode
-  final ThemeMode mode;
+  final int mode;
 
   // ignore: public_member_api_docs
   const InheritedThemeProvider(this.mode, {super.key, required super.child});